create lib folder and move all helper functions into lib. Fix eslint problems
