+++ .eslintrc
@@ -1,7 +1,7 @@
 {
   "root": true,
   "env": {
-    "node": false,
+    "node": true,
     "browser": true
   },
   "parserOptions": {
