+++ src/lib/check-for-semicolon.ts
@@ -1,0 +1,6 @@
+export default function checkForSemicolon(value: string): boolean {
+  if (!value) return false;
+  if (value.includes(';')) {
+    return true;
+  } else return false;
+}
