+++ src/lib/index.ts
@@ -1,0 +1,4 @@
+export { default as checkAnswer } from './check-answer';
+export { default as convertClassToReact } from './convert-class-to-react';
+export { default as checkForSemicolon } from './check-for-semicolon';
+export { default as removeSemiColon } from './remove-semicolon';
